+++ components/autofill/src/sync/credit_card/outgoing.rs
@@ -5,11 +5,12 @@
 
 use crate::db::models::credit_card::InternalCreditCard;
 use crate::db::schema::CREDIT_CARD_COMMON_COLS;
-use crate::encryption::EncryptorDecryptor;
+use crate::encryption::{decrypt_str, encrypt_str, EncryptorDecryptor};
 use crate::error::*;
 use crate::sync::common::*;
 use crate::sync::{credit_card::CreditCardPayload, OutgoingBso, ProcessOutgoingRecordImpl};
 use rusqlite::{Row, Transaction};
+use std::sync::Arc;
 use sync_guid::Guid as SyncGuid;
 
 const DATA_TABLE_NAME: &str = "credit_cards_data";
@@ -17,7 +18,7 @@ const MIRROR_TABLE_NAME: &str = "credit_cards_mirror";
 const STAGING_TABLE_NAME: &str = "credit_cards_sync_outgoing_staging";
 
 pub(super) struct OutgoingCreditCardsImpl {
-    pub(super) encdec: EncryptorDecryptor,
+    pub(super) encdec: Arc<dyn EncryptorDecryptor>,
 }
 
 impl ProcessOutgoingRecordImpl for OutgoingCreditCardsImpl {
@@ -47,12 +48,13 @@ impl ProcessOutgoingRecordImpl for OutgoingCreditCardsImpl {
             common_cols = CREDIT_CARD_COMMON_COLS,
         );
         let record_from_data_row: &dyn Fn(&Row<'_>) -> Result<(OutgoingBso, i64)> = &|row| {
-            let mut record = InternalCreditCard::from_row(row)?.into_payload(&self.encdec)?;
+            let mut record =
+                InternalCreditCard::from_row(row)?.into_payload(self.encdec.as_ref())?;
             // If the server had unknown fields we fetch it and add it to the record
             if let Some(enc_s) = row.get::<_, Option<String>>("payload")? {
                 // The full payload in the credit cards mirror is encrypted
                 let mirror_payload: CreditCardPayload =
-                    serde_json::from_str(&self.encdec.decrypt(&enc_s)?)?;
+                    serde_json::from_str(&decrypt_str(self.encdec.as_ref(), &enc_s)?)?;
                 record.entry.unknown_fields = mirror_payload.entry.unknown_fields;
             };
 
@@ -74,7 +76,7 @@ impl ProcessOutgoingRecordImpl for OutgoingCreditCardsImpl {
         .into_iter()
         .map(|(bso, change_counter)| {
             // Turn the record into an encrypted repr to save in the mirror.
-            let encrypted = self.encdec.encrypt(&bso.payload)?;
+            let encrypted = encrypt_str(self.encdec.as_ref(), &bso.payload)?;
             Ok((bso.envelope.id, encrypted, change_counter))
         })
         .collect::<Result<_>>()?;
@@ -110,6 +112,7 @@ impl ProcessOutgoingRecordImpl for OutgoingCreditCardsImpl {
 mod tests {
     use super::*;
     use crate::db::credit_cards::{add_internal_credit_card, tests::test_insert_mirror_record};
+    use crate::encryption::random_key_encryptor;
     use crate::sync::{common::tests::*, test::new_syncable_mem_db, UnknownFields};
     use serde_json::{json, Map, Value};
     use types::Timestamp;
@@ -164,7 +167,7 @@ mod tests {
             .clone()
     }
 
-    fn test_record(guid_prefix: char, encdec: &EncryptorDecryptor) -> InternalCreditCard {
+    fn test_record(guid_prefix: char, encdec: &dyn EncryptorDecryptor) -> InternalCreditCard {
         let json = test_json_record(guid_prefix);
         let payload = serde_json::from_value(json).unwrap();
         InternalCreditCard::from_payload(payload, encdec).expect("should be valid")
@@ -175,9 +178,9 @@ mod tests {
         let mut db = new_syncable_mem_db();
         let tx = db.transaction().expect("should get tx");
         let co = OutgoingCreditCardsImpl {
-            encdec: EncryptorDecryptor::new_with_random_key().unwrap(),
+            encdec: Arc::new(random_key_encryptor().unwrap()),
         };
-        let test_record = test_record('C', &co.encdec);
+        let test_record = test_record('C', co.encdec.as_ref());
 
         // create date record
         assert!(add_internal_credit_card(&tx, &test_record).is_ok());
@@ -196,9 +199,9 @@ mod tests {
         let mut db = new_syncable_mem_db();
         let tx = db.transaction().expect("should get tx");
         let co = OutgoingCreditCardsImpl {
-            encdec: EncryptorDecryptor::new_with_random_key().unwrap(),
+            encdec: Arc::new(random_key_encryptor().unwrap()),
         };
-        let test_record = test_record('C', &co.encdec);
+        let test_record = test_record('C', co.encdec.as_ref());
 
         // create tombstone record
         assert!(tx
@@ -231,19 +234,19 @@ mod tests {
         let mut db = new_syncable_mem_db();
         let tx = db.transaction().expect("should get tx");
         let co = OutgoingCreditCardsImpl {
-            encdec: EncryptorDecryptor::new_with_random_key().unwrap(),
+            encdec: Arc::new(random_key_encryptor().unwrap()),
         };
 
         // create synced record with non-zero sync_change_counter
-        let mut test_record = test_record('C', &co.encdec);
+        let mut test_record = test_record('C', co.encdec.as_ref());
         let initial_change_counter_val = 2;
         test_record.metadata.sync_change_counter = initial_change_counter_val;
         assert!(add_internal_credit_card(&tx, &test_record).is_ok());
         let guid = test_record.guid.clone();
         //test_insert_mirror_record doesn't encrypt the mirror payload, but in reality we do
         // so we encrypt here so our fetch_outgoing_records doesn't break
-        let mut bso = test_record.into_test_incoming_bso(&co.encdec, Default::default());
-        bso.payload = co.encdec.encrypt(&bso.payload).unwrap();
+        let mut bso = test_record.into_test_incoming_bso(co.encdec.as_ref(), Default::default());
+        bso.payload = encrypt_str(co.encdec.as_ref(), &bso.payload).unwrap();
         test_insert_mirror_record(&tx, bso);
         exists_with_counter_value_in_table(&tx, DATA_TABLE_NAME, &guid, initial_change_counter_val);
 
@@ -262,16 +265,16 @@ mod tests {
         let mut db = new_syncable_mem_db();
         let tx = db.transaction().expect("should get tx");
         let co = OutgoingCreditCardsImpl {
-            encdec: EncryptorDecryptor::new_with_random_key().unwrap(),
+            encdec: Arc::new(random_key_encryptor().unwrap()),
         };
 
         // create synced record with no changes (sync_change_counter = 0)
-        let test_record = test_record('C', &co.encdec);
+        let test_record = test_record('C', co.encdec.as_ref());
         let guid = test_record.guid.clone();
         assert!(add_internal_credit_card(&tx, &test_record).is_ok());
         test_insert_mirror_record(
             &tx,
-            test_record.into_test_incoming_bso(&co.encdec, Default::default()),
+            test_record.into_test_incoming_bso(co.encdec.as_ref(), Default::default()),
         );
 
         do_test_outgoing_synced_with_no_change(
@@ -288,11 +291,11 @@ mod tests {
         let mut db = new_syncable_mem_db();
         let tx = db.transaction().expect("should get tx");
         let co = OutgoingCreditCardsImpl {
-            encdec: EncryptorDecryptor::new_with_random_key().unwrap(),
+            encdec: Arc::new(random_key_encryptor().unwrap()),
         };
 
         // create synced record with non-zero sync_change_counter
-        let mut test_record = test_record('D', &co.encdec);
+        let mut test_record = test_record('D', co.encdec.as_ref());
         let initial_change_counter_val = 2;
         test_record.metadata.sync_change_counter = initial_change_counter_val;
         assert!(add_internal_credit_card(&tx, &test_record).is_ok());
@@ -304,8 +307,8 @@ mod tests {
         // so we encrypt here so our fetch_outgoing_records doesn't break
         let mut bso = test_record
             .clone()
-            .into_test_incoming_bso(&co.encdec, unknown_fields);
-        bso.payload = co.encdec.encrypt(&bso.payload).unwrap();
+            .into_test_incoming_bso(co.encdec.as_ref(), unknown_fields);
+        bso.payload = encrypt_str(co.encdec.as_ref(), &bso.payload).unwrap();
         test_insert_mirror_record(&tx, bso);
         exists_with_counter_value_in_table(
             &tx,